requirements.txt
einops,
hydra-core>=1.3.2
mne>=1.7.1
termcolor
torch>=2.0.1
torchaudio
torchmetrics>=0.11.4
torchvision
wandb>=0.15.0
ipykernel>=6.29.5
torchsummary>=1.5.1
ipywidgets>=8.1.3
pyvistaqt>=0.11.1
ipympl>=0.9.4
pyvista>=0.44.0
ipyvtklink>=0.2.3